Revert "packages moved, FPS printed in window title"
This reverts commit 249c5b199d6ff2d49dce818b2a035e04c26ca549.
