+++ src/main/java/fr/utbm/info/vi51/worldswar/gui/Messages.java
@@ -1,4 +1,4 @@
-package fr.utbm.info.vi51.worldswar.gui.messages;
+package fr.utbm.info.vi51.worldswar.gui;
 
 import java.util.MissingResourceException;
 import java.util.ResourceBundle;
@@ -10,7 +10,7 @@ import java.util.ResourceBundle;
  *         messages
  */
 public class Messages {
-	private static final String BUNDLE_NAME = "fr.utbm.info.vi51.worldswar.gui.messages.messages"; //$NON-NLS-1$
+	private static final String BUNDLE_NAME = "fr.utbm.info.vi51.worldswar.gui.messages"; //$NON-NLS-1$
 
 	private static final ResourceBundle RESOURCE_BUNDLE = ResourceBundle.getBundle(BUNDLE_NAME);
